refactor: countdown timer useEffect
